+++ src/EntryProvider.php
@@ -16,4 +16,10 @@ interface EntryProvider
      * @return static The initialized entry provider
      */
     public static function initialize(): self;
+
+    /**
+     * Returns list of provided container entries and associated methods.
+     * @return string[] List of provided container entries and associated methods
+     */
+    public function getMethods(): array;
 }