+++ tests/tests/Entry/ProviderEntryTest.php
@@ -3,6 +3,10 @@
 namespace Simply\Container\Entry;
 
 use PHPUnit\Framework\TestCase;
+use Simply\Container\AbstractEntryProvider;
+use Simply\Container\Container;
+use Simply\Container\ContainerBuilder;
+use Simply\Container\Exception\ContainerException;
 
 /**
  * ProviderEntryTest.
@@ -15,8 +19,21 @@ class ProviderEntryTest extends TestCase
     public function testInvalidCallableFormat()
     {
         $this->expectException(\InvalidArgumentException::class);
+
         new ProviderEntry(function () {
             return 'foobar';
         });
+    }
+
+    public function testUnexpectedProviderValue()
+    {
+        $container = new Container();
+        $date = new \DateTime();
+
+        $container->addEntry('timestamp', new ProviderEntry([$date, 'getTimestamp']));
+        $container->addEntry(\DateTime::class, new MixedEntry('Not A Object'));
+
+        $this->expectException(ContainerException::class);
+        $container->get('timestamp');
     }
 }