+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Moved the logic of detecting provider methods to new interface method `EntryProvider::getMethods()`.
+- Prefer constructors in cache loading, when possible
+
+### Fixed
+- `ProviderEntry` will now ensure the value returned by container is actually an object
 
 ## 0.1.0 - 2018-07-06
 ### Added